pyproject.toml
[tool.poetry]
name = "bitrix-gptbot"
version = "0.1.0"
description = ""
authors = ["Kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
freegpt = "^1.3.4"
asyncpg = "^0.29.0"
celery = "^5.3.6"
redis = "^5.0.2"
types-requests = "^2.31.0.20240311"

[tool.poetry.group.lint.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
